+++ src/crypto/test/key_exchange.cpp
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 
-#include "../key_exchange.h"
+#include "crypto/key_exchange.h"
 
 #include <doctest/doctest.h>
